:truck: Move notification channel tests into notifications app
